pyproject.toml
[tool.poetry]
name = "recipe-recommender"
version = "0.1.0"
description = ""
authors = ["Bucky Roberts <buckyroberts@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
openai = "0.28"
llama-index = "^0.8.29.post1"
promptlayer = "^1.0.26"
pinecone-client = "^5.0.1"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
